pyproject.toml
[tool.poetry]
name = "gea-mnemon-reproduce"
version = "0.1.0"
description = ""
authors = ["kjchan"]
readme = "README.md"
packages = [{include = "gea_mnemon_reproduce"}]

[tool.poetry.dependencies]
python = "^3.11"
torch-geometric = "^2.3.1"
torch = "^2.0.1"
torchvision = "^0.15.2"
torchaudio = "^2.0.2"
python-dotenv = "^1.0.0"
pre-commit = "^3.4.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
